feat: add 'run' command for executing FerroPhase files and enhance eval command with pipeline integration

+++ crates/fp-cli/src/commands/compile.rs
@@ -126,7 +126,7 @@ async fn compile_file(
     
     // Configure pipeline for compilation
     let pipeline_config = PipelineConfig {
-        optimization_level: args.opt_level as u32,
+        optimization_level: args.opt_level as u8,
         print_ast: false,
         print_passes: false,
         target: args.target.clone(),
